Convert alsa event time to milliseconds
Also, make time field in MidiEvent long long to gurantee 64 bit size.

+++ include/sigmidi.h
@@ -15,7 +15,7 @@ struct MidiEvent {
     snd_seq_event_type_t type;
     unsigned char note;
     unsigned char velocity;
-    long time;
+    long long time;
 };
 
 struct Note {
